// src/pages/Licenciado/modules/EditRegistrationLA/components/interface.ts
export interface SellerData {
  trading_name: string;
  document: string;
  type_document: string;
  type: string;
  email: string;
  status: string;
  company_name: string;
  opening_date: string;
  mcc: string;
  phone: string;
  owner_name: string;
  owner_birthday: string;
  owner_cpf: string;
  owner_rg: string;
  address_cep: string;
  address_street: string;
  address_number: number;
  address_complement: string;
  address_neighborhood: string;
  address_city: string;
  address_state: string;
  tpv_goal: number;
}